@@ -5,7 +5,7 @@ author = Doug Hellmann
 author-email = doug@doughellmann.com
 summary = Sphinx spelling extension
 home-page = https://github.com/sphinx-contrib/spelling
-python_requires = >=3.5
+python_requires = >=3.6
 classifier =
     Development Status :: 5 - Production/Stable
     Environment :: Console
@@ -17,7 +17,6 @@ classifier =
     Programming Language :: Python
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
